Split menu code into tabs
