+++ lib/menu/tabs/main/artists.tsx
